DB-2337: - Create getPreview helper - Implement getPreview for content in the starter (recipes, pages, articles)
